Add in support for Uploading Account Images
Added the ability for Veterans to have profile pictures, and for
Businesses to have both profile and cover images. The servlet
UploadImage is solely responsible for uploading temporary images
for both profile and cover images. To make sure that the images
correspond to the sites styles, the photos are properly scaled depending
on whether the user is uploading a profile or cover picture.

Photos uploaded with the UploadImage servlet are not quite ready
to be set as the account images. Once they are uploaded, the user
will need to crop any profile pictures, and reposition any cover
images. This is done through another servlet, ResizePhoto.

This change also required changes to the database schema. MySQL
was modified to allow Veterans and Businesses to update their
profiles with images to display. Also, while scrolling through,
several comments were updated and added.