Rework User registration and update functionality
Reworked the User registration script and update functionality in the
following ways.

A new servlet called UpdateBusiness was added. This is responsible for
taking a post request, extracting all the keys, and updating the
business with all keys from the request that exist in the database
table. To help with this process, a method UpdateBusiness was also
added to the MySQL class, which takes in a map of string and objects,
and creates a PreparedStatement designed for updating a Business
profile.

To help with Key validation, an array of database columns was added
to the Business object. To determine if a particular key is apart
of a database object, a new method was added to the Utilities class
that takes in a string and an array of strings, and determines if
the given string is apart of the given array.

To rework the Business Registration, all the functionality that was
related to creating a user account was removed from that class
and put in it's own class UserRegistration. This contains static
methods for checking to make sure user input is not null for all
required fields, and that phone numbers are emails are of the proper input
type.

Also, a few functions dealing with profile and cover images that were left
out of the previous commit were added here. Functions looking for a temp
uploaded file in the database were removed, as the src of the photos
just appear in the account tables.