Add in Various files for Registration.
Added in different styles and photos associated with the styles.

Added a login page that looks for a valid account from a given
email and password. It determines what type of account it is by
checking all the tables for that same user id.

Modified the editBusiness, MySQL, and VeteranRegistration pages to
correspond to the new Registration process.