Added in various library files.
Added in various jQuery files for dynamic forms, submitting forms,
providing input masts, and draggable objects.

Added in various custom objects.

CoverImageDrag allows an image to be dragged up and down for
repositioning.

CropPhoto allows a photo to be cropped using a draggable rectangle.

LightBox places a div that fill the entire screen with a light black
transparent background with a div in the middle. This class is designed
to hold profile photos for users to crop, but can be used for other
objects as well.

Added in the MimeTypes class, which is a list of different
possible mime types for an uploaded file.