Modify the User registration process
Added the Enum to the Business class that corresponds to the database
columns associated to a Business object. This helps validate keys with the
UpdateBusiness class.

A few minor changes were made as well. UserRegistration uses a generated
password script to create passwords for users as opposed to having the
user set them. Users can still set passwords, just it would be on their
profile edit page after they created their account. Also, a SQL
error was fixed in the MySQL class.