Updates to the Password Validator
Modified the password validator so that it uses actual strings
instead of regular expressions to check for the number of
alphabetical characters and digits.

All the validations were moved from if statements to associated
validation functions.

Minimun password length and number of special characters were removed
from the validator, as they are no longer used.