Fix Various Registration Bugs
Made sure the redirect for the UpdateUserProfile page
redirects based on whether the business id is st as opposed to
the user id.

Veteran Registration redirects to the veteran page if any
registration errors are found.

Removed an unused import from the User class.

Renamed column that was incorrectly named in the Veteran class, and
added a function to validate that an incoming key and value is
a valid input for the Veteran database table.