Modify MySQL class to properly update user profile information related to photos and account specific (business, veteran) information.
Added support for Questions to the MySQL class. This required adding
a Question class to the project.

Added support to get the tags from a given account, and to get the
businesses associated with a given input of tags.

Added the question object which represents a question that a veteran
can ask the enterpreneurs.