Add in mailing class
Added in a Java class that allows for sending emails through the
server. Right now, the only support is for sending basic messages
to users. There are a few things about this class that need to be
addressed.

The password is provided in the email I sent. Make sure you copy
the password I send you to the email class before you use it.

The email address provided is temporary. We will need to switch
once we have a good emailing system in place.