Make base soldier up account class.
Made a base class for Veterans and Business to extend off of. This
allows us to keep the common properties between the two classes
in the same location.

Also, the database was updated to have an Accounts table to store
all the common properties. The queries in the MySQL class were updated
as well.

In addition to changing the account queries, the MySQL class was given generic
table functions for updating and deleting from tables. Formatting such
as trailing whitespace and unnecessary casts were removed as well.

Since we have a unified account type now, the vid and bid attributes
were changed for both to be aid. To determine what kind of account the
aid refers to, the session attribute editing_account_type can be used.

Note: To use this commit, make sure you use the modified sql script
in this commit.