Modify LightBox to use more than images
Modified the LightBox so that it does not require an image. This
allows us to use the LightBox for objects such as forms. To do this
different styles were changed with the divs so that the main box
is centered on the screen, and it's height and width is not based
on a passed in image.