Add tags to veteran questions
Added the ability for questions to contain tags. The VeteranQuestionForm
contains three different autocomplete tags that search the tags table.

To make these fields work, the MySQL was updated to include questions
for attaching tags to questions, and also removing them.

Since we query tags for more than accounts now, the original
getSimiliarTags was changed so that it does not query on the users
current account. Another function getSimiliarTagsNotInAccount was
added that searches for non existing tags in a user account. The
TagSubscriber was updated as well to use this new functions.

The QueryTags servlet was changed so that it can take an input parameter
of whether we are querying for an account. The TagScripts was changed
to use this new parameter as well. It calls the appropriate function
depending on that new parameter.