Fix various presentation and build bugs
Renamed the href in various jsp pages for the main styles page from
Styles to styles, as the web browser is case sensitive.

The session attribute of id was changed to uid, and the bid and vid
attributes were removed to be the generic account id (aid). To identify
the account type, another session attribute editing_account_type was added
to the project. Several different classes were changed so that they use
these new attributes as opposed to the old ones.

The page editVeteranProfile had the action and form input names changed
so that it can be used with the Login servlet.

The MySQL class has the getPreparedStatement function moved. Several
functions were changed to just get the prepared statement from the
database connection. Also, the get businesses function was changed to
an array list, and several different query bugs were fixed.
that dealt with renamed session attributes.

The Divisions class was deleted as it is not currently used by any of
the other files in the project.