Modify Header to use generic account types
Modified the header include so that checks the aid for a logged in
account instead of either a vid or bid. To determine what account
is being used, the editing_account_type session variable is checked.

The Login servlet was also changed to use generic account types as
opposed to a bid or vid session variable.