Fix build issues with branch
Fixed various build issues with the mongodb branch. Issues remained
after the master branch was initially merged into this branch, and this
commit fixes several of the things that were left out or overlooked.

A few of the classes needed long id's added to them in addition to
the mongodb id. This is just to get the site to build with the MySQL
database.

Also, several unneeded files were removed from the project.