Use Engine for all database queries
Modified the website so that all queries go though the Engine class.
To do this, all methods in the MySQL class were added to the Engine
class.

Once the engine class was done, all MySQL instances in the project
were replaced with Engine.

This commit fixed a few other things as well. Summaries only exist
with businesses, so those variables were moved from the SoldierUpAccount
to the Business class. Veteran needed a numeric id for mysql storage, so one
was added. This will most likely be removed later.